Confirm card delete and go Back button

+++ src/pages/EditQuestionPage/EditQuestion.jsx
@@ -4,7 +4,7 @@ import QuestionForm from '../../components/QuestionForm';
 import MarkdownRenderer from '../../components/MarkdownRenderer';
 import { dateFormat } from '../../helpers/dateFormat';
 import { useFetch } from '../../hooks/useFetch';
-import { Info, X } from 'lucide-react';
+import { Info, X, ArrowLeft } from 'lucide-react';
 import { API_URL } from '../../constants';
 
 import styles from './EditQuestionPage.module.css';
@@ -67,6 +67,12 @@ const EditQuestion = ({ initialState = {} }) => {
 	};
 
 	const [removeQuestion, isQuestionRemoving] = useFetch(async () => {
+		let check = confirm('Are your sure you want to delete this card?');
+
+		if (!check) {
+			return;
+		}
+
 		await fetch(`${API_URL}/checkycards/${initialState.id}`, {
 			method: 'DELETE',
 		});
@@ -78,7 +84,12 @@ const EditQuestion = ({ initialState = {} }) => {
 		<>
 			<div className={styles.formContainer}>
 				<div className={styles.topContainer}>
-					<h1 className={styles.formTitle}>Edit card</h1>
+					<div className={styles.topContainerLeft}>
+						<button className={styles.goBackBtn} onClick={() => navigate((-1))} disabled={isQuestionRemoving}>
+							<ArrowLeft />
+						</button>
+						<h1 className={styles.formTitle}>Edit card</h1>
+					</div>
 					<div className={styles.topContainerRight}>
 						<button className={styles.infoBtn} onClick={() => setIsModalOpen(true)}>
 							<Info />
@@ -114,7 +125,6 @@ const EditQuestion = ({ initialState = {} }) => {
 								</p>
 							</div>
 						</div>
-						
 
 						<div className={`${styles.example} ${styles.section}`}>
 							<div className={styles.exampleBlock}>
